Move widgets and adv_widgets to render module
